+++ glide-core/src/request_type.rs
@@ -151,7 +151,10 @@ pub enum RequestType {
     ZRandMember = 139,
     Bitcount = 140,
     BZMPop = 141,
-    ZInter = 144,
+    SetBit = 142,
+    ZInterCard = 143,
+    ZMPop = 144,
+    ZInter = 145,
 }
 
 fn get_two_word_command(first: &str, second: &str) -> Cmd {
@@ -306,6 +309,9 @@ impl From<::protobuf::EnumOrUnknown<ProtobufRequestType>> for RequestType {
             ProtobufRequestType::ZRandMember => RequestType::ZRandMember,
             ProtobufRequestType::Bitcount => RequestType::Bitcount,
             ProtobufRequestType::BZMPop => RequestType::BZMPop,
+            ProtobufRequestType::SetBit => RequestType::SetBit,
+            ProtobufRequestType::ZInterCard => RequestType::ZInterCard,
+            ProtobufRequestType::ZMPop => RequestType::ZMPop,
             ProtobufRequestType::ZInter => RequestType::ZInter,
         }
     }
@@ -457,6 +463,9 @@ impl RequestType {
             RequestType::ZRandMember => Some(cmd("ZRANDMEMBER")),
             RequestType::Bitcount => Some(cmd("BITCOUNT")),
             RequestType::BZMPop => Some(cmd("BZMPOP")),
+            RequestType::SetBit => Some(cmd("SETBIT")),
+            RequestType::ZInterCard => Some(cmd("ZINTERCARD")),
+            RequestType::ZMPop => Some(cmd("ZMPOP")),
             RequestType::ZInter => Some(cmd("ZINTER")),
         }
     }